Replace generic errors, pep8, type() -> isinstance()
